Add patient unit tests broken down into multiple files to recude execution time
