+++ frontend/src/app/patients/AddPatient.test.tsx
@@ -1,149 +1,18 @@
-import { render, screen, within, waitFor } from "@testing-library/react";
-import { MockedProvider } from "@apollo/client/testing";
-import { Provider } from "react-redux";
-import configureStore from "redux-mock-store";
-import { MemoryRouter, Route, Routes, useLocation } from "react-router-dom";
-import userEvent, { UserEvent } from "@testing-library/user-event";
+import { screen } from "@testing-library/react";
 
 import * as smartyStreets from "../utils/smartyStreets";
-import SRToastContainer from "../commonComponents/SRToastContainer";
 import { PATIENT_TERM } from "../../config/constants";
 
-import AddPatient, { ADD_PATIENT, PATIENT_EXISTS } from "./AddPatient";
-
-interface LocationOptions {
-  search: string;
-  pathname: string;
-  state: {
-    patientId: string;
-  };
-}
-
-const mockFacilityID = "b0d2041f-93c9-4192-b19a-dd99c0044a7e";
-const mockStore = configureStore([]);
-const store = mockStore({
-  facilities: [{ id: mockFacilityID, name: "123" }],
-  organization: { name: "Test Organization" },
-});
-
-const RouterWithFacility: React.FC<RouterWithFacilityProps> = ({
-  children,
-}) => (
-  <MemoryRouter initialEntries={[`/add-patient?facility=${mockFacilityID}`]}>
-    <Routes>{children}</Routes>
-  </MemoryRouter>
-);
-
-const fillOutForm = async (
-  user: UserEvent,
-  inputs: { [label: string]: string },
-  dropdowns: { [label: string]: string },
-  inputGroups: {
-    [legend: string]: { label: string; value: string; exact?: boolean };
-  }
-) => {
-  const inputElements = Object.entries(inputs);
-
-  for (let i = 0; i < inputElements.length; i++) {
-    await user.type(
-      screen.getByLabelText(inputElements[i][0], {
-        exact: false,
-      }),
-      inputElements[i][1]
-    );
-  }
-
-  const dropDownElements = Object.entries(dropdowns);
-
-  for (let i = 0; i < dropDownElements.length; i++) {
-    await user.selectOptions(
-      screen.getByLabelText(dropDownElements[i][0], {
-        exact: false,
-      }),
-      dropDownElements[i][1]
-    );
-  }
-
-  const inputGroupsElements = Object.entries(inputGroups);
-
-  for (let i = 0; i < inputGroupsElements.length; i++) {
-    const [legend, { label, exact }] = inputGroupsElements[i];
-    const fieldset = screen
-      .getByText(legend, {
-        exact: true,
-      })
-      .closest("fieldset");
-    if (fieldset === null) {
-      throw Error(`Unable to corresponding fieldset for ${legend}`);
-    }
-    await user.click(
-      within(fieldset).getByLabelText(label, {
-        exact: exact || false,
-      })
-    );
-  }
-};
-
-const addPatientRequestParams = {
-  firstName: "Alice",
-  middleName: null,
-  lastName: "Hamilton",
-  lookupId: null,
-  birthDate: "1970-09-22",
-  street: "25 Shattuck St",
-  streetTwo: null,
-  city: "Boston",
-  state: "MA",
-  zipCode: "02115",
-  country: "USA",
-  telephone: null,
-  phoneNumbers: [
-    {
-      type: "MOBILE",
-      number: "617-432-1000",
-    },
-  ],
-  role: null,
-  emails: ["foo@bar.org"],
-  county: "",
-  race: "other",
-  ethnicity: "refused",
-  gender: "female",
-  genderIdentity: "female",
-  facilityId: mockFacilityID,
-  preferredLanguage: null,
-  testResultDelivery: "SMS",
-};
-
-const addPatientRequestNoDelivery = {
-  ...addPatientRequestParams,
-  testResultDelivery: null,
-};
-
-const addPatientRequestNoAddressValidation = {
-  ...addPatientRequestParams,
-  county: null,
-};
+import {
+  mockFacilityID,
+  renderWithUserNoFacility,
+  renderWithUserWithFacility,
+} from "./AddPatientTestUtils";
 
 describe("AddPatient", () => {
   describe("No facility selected", () => {
-    const renderWithUser = () => ({
-      user: userEvent.setup(),
-      ...render(
-        <Provider store={store}>
-          <MockedProvider mocks={[]} addTypename={false}>
-            <MemoryRouter>
-              <Routes>
-                <Route path="/" element={<AddPatient />} />
-              </Routes>
-            </MemoryRouter>
-          </MockedProvider>
-        </Provider>
-      ),
-    });
-
     it("does not show the form title", () => {
-      renderWithUser();
+      renderWithUserNoFacility();
       expect(
         screen.queryByText(`Add new ${PATIENT_TERM}`, {
           exact: false,
@@ -151,7 +20,7 @@ describe("AddPatient", () => {
       ).not.toBeInTheDocument();
     });
     it("shows a 'No facility selected' message", async () => {
-      renderWithUser();
+      renderWithUserNoFacility();
       expect(
         screen.getByText("No facility selected", {
           exact: false,
@@ -161,129 +30,16 @@ describe("AddPatient", () => {
   });
 
   describe("happy path", () => {
-    let zipCodeSpy: jest.SpyInstance;
-
     beforeEach(async () => {
       jest.spyOn(smartyStreets, "getBestSuggestion").mockImplementation();
       jest
         .spyOn(smartyStreets, "suggestionIsCloseEnough")
         .mockReturnValue(false);
       jest.spyOn(smartyStreets, "getZipCodeData").mockResolvedValue(undefined);
-      zipCodeSpy = jest
-        .spyOn(smartyStreets, "isValidZipCodeForState")
-        .mockReturnValue(true);
-    });
-    const mocks = [
-      {
-        request: {
-          query: PATIENT_EXISTS,
-          variables: {
-            firstName: "Alice",
-            lastName: "Hamilton",
-            birthDate: "1970-09-22",
-            facilityId: "b0d2041f-93c9-4192-b19a-dd99c0044a7e",
-          },
-        },
-        result: {
-          data: {
-            patientExistsWithoutZip: false,
-          },
-        },
-      },
-      {
-        request: {
-          query: PATIENT_EXISTS,
-          variables: {
-            firstName: "Alice",
-            lastName: "Hamilton",
-            birthDate: "1970-09-22",
-            facilityId: "",
-          },
-        },
-        result: {
-          data: {
-            patientExistsWithoutZip: false,
-          },
-        },
-      },
-      {
-        request: {
-          query: ADD_PATIENT,
-          variables: addPatientRequestParams,
-        },
-        result: {
-          data: {
-            addPatient: {
-              internalId: "153f661f-b6ea-4711-b9ab-487b95198cce",
-              facility: {
-                id: "facility-id-001",
-              },
-            },
-          },
-        },
-      },
-      {
-        request: {
-          query: ADD_PATIENT,
-          variables: addPatientRequestNoDelivery,
-        },
-        result: {
-          data: {
-            addPatient: {
-              internalId: "153f661f-b6ea-4711-b9ab-487b95198cce",
-              facility: {
-                id: "facility-id-001",
-              },
-            },
-          },
-        },
-      },
-      {
-        request: {
-          query: ADD_PATIENT,
-          variables: addPatientRequestNoAddressValidation,
-        },
-        result: {
-          data: {
-            addPatient: {
-              internalId: "153f661f-b6ea-4711-b9ab-487b95198cce",
-              facility: {
-                id: "facility-id-001",
-              },
-            },
-          },
-        },
-      },
-    ];
-    const Queue = () => {
-      const location = useLocation() as LocationOptions;
-
-      return (
-        <p>
-          Testing Queue! {location.search} {location.state.patientId}
-        </p>
-      );
-    };
-    const renderWithUser = () => ({
-      user: userEvent.setup(),
-      ...render(
-        <>
-          <Provider store={store}>
-            <MockedProvider mocks={mocks} addTypename={false}>
-              <RouterWithFacility>
-                <Route element={<AddPatient />} path={"/add-patient"} />
-                <Route path={"/patients"} element={<p>Patients!</p>} />
-                <Route path={"/queue"} element={<Queue />} />
-              </RouterWithFacility>
-            </MockedProvider>
-          </Provider>
-          <SRToastContainer />
-        </>
-      ),
     });
 
     it("shows the form title", async () => {
-      renderWithUser();
+      renderWithUserWithFacility();
       expect(
         (
           await screen.findAllByText(`Add new ${PATIENT_TERM}`, {
@@ -295,7 +51,7 @@ describe("AddPatient", () => {
 
     describe("Choosing a country", () => {
       it("should show the state and zip code inputs for USA", async () => {
-        const { user } = renderWithUser();
+        const { user } = renderWithUserWithFacility();
 
         await user.selectOptions(
           screen.getByLabelText("Country", { exact: false }),
@@ -305,7 +61,7 @@ describe("AddPatient", () => {
         expect(await screen.findByText("ZIP code")).toBeInTheDocument();
       });
       it("should show the state and zip code inputs for Canada", async () => {
-        const { user } = renderWithUser();
+        const { user } = renderWithUserWithFacility();
         await user.selectOptions(
           screen.getByLabelText("Country", { exact: false }),
           "CAN"
@@ -314,7 +70,7 @@ describe("AddPatient", () => {
         expect(await screen.findByText("ZIP code")).toBeInTheDocument();
       });
       it("should show different states for Canada", async () => {
-        const { user } = renderWithUser();
+        const { user } = renderWithUserWithFacility();
 
         await user.selectOptions(
           screen.getByLabelText("Country", { exact: false }),
@@ -329,8 +85,9 @@ describe("AddPatient", () => {
         await user.selectOptions(stateInput, "QC");
         expect(stateInput.value).toBe("QC");
       });
+
       it("should hide the state and zip code inputs for non-US countries", async () => {
-        const { user } = renderWithUser();
+        const { user } = renderWithUserWithFacility();
 
         await user.selectOptions(
           screen.getByLabelText("Country", { exact: false }),
@@ -341,177 +98,9 @@ describe("AddPatient", () => {
       });
     });
 
-    describe("All required fields entered and submitting address verification", () => {
-      it("redirects to the person tab", async () => {
-        const { user } = renderWithUser();
-
-        await fillOutForm(
-          user,
-          {
-            "First Name": "Alice",
-            "Last Name": "Hamilton",
-            "Date of birth": "1970-09-22",
-            "Primary phone number": "617-432-1000",
-            "Email address": "foo@bar.org",
-            "Street address 1": "25 Shattuck St",
-            City: "Boston",
-
-            "ZIP code": "02115",
-          },
-          { Facility: mockFacilityID, State: "MA", Country: "USA" },
-          {
-            "Phone type": {
-              label: "Mobile",
-              value: "MOBILE",
-              exact: true,
-            },
-            "Would you like to receive your results via text message?": {
-              label: "Yes",
-              value: "SMS",
-              exact: false,
-            },
-            Race: {
-              label: "Other",
-              value: "other",
-              exact: true,
-            },
-            "Are you Hispanic or Latino?": {
-              label: "Prefer not to answer",
-              value: "refused",
-              exact: true,
-            },
-            "Sex assigned at birth": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-            "What's your gender identity?": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-          }
-        );
-
-        await user.click(screen.queryAllByText(/Save Changes/i)[0]);
-
-        await screen.findByText(/Address validation/i);
-        const modal = screen.getByRole("dialog");
-
-        await user.click(
-          within(modal).getByLabelText("Use address as entered", {
-            exact: false,
-          })
-        );
-
-        await user.click(
-          within(modal).getByText("Save changes", {
-            exact: false,
-          })
-        );
-        expect(await screen.findByText("Patients!")).toBeInTheDocument();
-      });
-      it("surfaces an error if invalid zip code for state", async () => {
-        const { user } = renderWithUser();
-        zipCodeSpy.mockReturnValue(false);
-        await fillOutForm(
-          user,
-          {
-            "First Name": "Alice",
-            "Last Name": "Hamilton",
-            "Date of birth": "1970-09-22",
-            "Primary phone number": "617-432-1000",
-            "Email address": "foo@bar.org",
-            "Street address 1": "25 Shattuck St",
-            City: "Boston",
-            "ZIP code": "02115",
-          },
-          { Facility: mockFacilityID, State: "MA", Country: "USA" },
-          {
-            "Phone type": {
-              label: "Mobile",
-              value: "MOBILE",
-              exact: true,
-            },
-            "Would you like to receive your results via text message?": {
-              label: "Yes",
-              value: "SMS",
-              exact: false,
-            },
-            Race: {
-              label: "Other",
-              value: "other",
-              exact: true,
-            },
-            "Are you Hispanic or Latino?": {
-              label: "Prefer not to answer",
-              value: "refused",
-              exact: true,
-            },
-            "Sex assigned at birth": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-          }
-        );
-        await user.click(
-          screen.queryAllByText("Save Changes", {
-            exact: false,
-          })[0]
-        );
-        await screen.findByText(/Invalid ZIP code for this state/i);
-      });
-      it("requires race field to be populated", async () => {
-        const { user } = renderWithUser();
-        await fillOutForm(
-          user,
-          {
-            "First Name": "Alice",
-            "Last Name": "Hamilton",
-            "Date of birth": "1970-09-22",
-            "Primary phone number": "617-432-1000",
-            "Email address": "foo@bar.org",
-            "Street address 1": "25 Shattuck St",
-            City: "Boston",
-
-            "ZIP code": "02115",
-          },
-          { Facility: mockFacilityID, State: "MA", Country: "USA" },
-          {
-            "Phone type": {
-              label: "Mobile",
-              value: "MOBILE",
-              exact: true,
-            },
-            "Would you like to receive your results via text message?": {
-              label: "Yes",
-              value: "SMS",
-              exact: false,
-            },
-            "Sex assigned at birth": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-          }
-        );
-
-        await user.click(
-          screen.queryAllByText("Save Changes", {
-            exact: false,
-          })[0]
-        );
-
-        await waitFor(() =>
-          expect(screen.queryByText("Error: Race is missing"))
-        );
-      });
-    });
-
     describe("facility select input", () => {
       it("is present in the form", () => {
-        renderWithUser();
+        renderWithUserWithFacility();
         const facilityInput = screen.getByLabelText("Facility", {
           exact: false,
         }) as HTMLSelectElement;
@@ -519,7 +108,7 @@ describe("AddPatient", () => {
       });
 
       it("defaults to no selection", () => {
-        renderWithUser();
+        renderWithUserWithFacility();
         const facilityInput = screen.getByLabelText("Facility", {
           exact: false,
         }) as HTMLSelectElement;
@@ -527,7 +116,7 @@ describe("AddPatient", () => {
       });
 
       it("updates its selection on change", async () => {
-        const { user } = renderWithUser();
+        const { user } = renderWithUserWithFacility();
         const facilityInput = screen.getByLabelText("Facility", {
           exact: false,
         }) as HTMLSelectElement;
@@ -538,284 +127,10 @@ describe("AddPatient", () => {
 
     describe("With student ID", () => {
       it("allows student ID to be entered", async () => {
-        const { user } = renderWithUser();
+        const { user } = renderWithUserWithFacility();
         await user.selectOptions(screen.getByLabelText("Role"), "STUDENT");
         expect(await screen.findByText("Student ID")).toBeInTheDocument();
       });
-    });
-
-    describe("saving changes and starting a test", () => {
-      it("redirects to the queue after address validation", async () => {
-        const { user } = renderWithUser();
-        await fillOutForm(
-          user,
-          {
-            "First Name": "Alice",
-            "Last Name": "Hamilton",
-            "Date of birth": "1970-09-22",
-            "Primary phone number": "617-432-1000",
-            "Email address": "foo@bar.org",
-            "Street address 1": "25 Shattuck St",
-            City: "Boston",
-
-            "ZIP code": "02115",
-          },
-          { Facility: mockFacilityID, State: "MA", Country: "USA" },
-          {
-            "Phone type": {
-              label: "Mobile",
-              value: "MOBILE",
-              exact: true,
-            },
-            "Would you like to receive your results via text message?": {
-              label: "Yes",
-              value: "SMS",
-              exact: false,
-            },
-            Race: {
-              label: "Other",
-              value: "other",
-              exact: true,
-            },
-            "Are you Hispanic or Latino?": {
-              label: "Prefer not to answer",
-              value: "refused",
-              exact: true,
-            },
-            "Sex assigned at birth": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-            "What's your gender identity?": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-          }
-        );
-        jest
-          .spyOn(smartyStreets, "suggestionIsCloseEnough")
-          .mockReturnValue(true);
-
-        await user.click(
-          screen.queryAllByText("Save and start test", {
-            exact: false,
-          })[0]
-        );
-
-        //await new Promise((resolve) => setTimeout(resolve, 0));
-
-        expect(
-          screen.queryByText("Address validation", {
-            exact: false,
-          })
-        ).not.toBeInTheDocument();
-
-        await screen.findByText(/Testing Queue!/i);
-      });
-
-      it("redirects to the queue with a patient id and selected facility id", async () => {
-        const { user } = renderWithUser();
-        await fillOutForm(
-          user,
-          {
-            "First Name": "Alice",
-            "Last Name": "Hamilton",
-            "Date of birth": "1970-09-22",
-            "Primary phone number": "617-432-1000",
-            "Email address": "foo@bar.org",
-            "Street address 1": "25 Shattuck St",
-            City: "Boston",
-            "ZIP code": "02115",
-          },
-          { Facility: mockFacilityID, State: "MA", Country: "USA" },
-          {
-            "Phone type": {
-              label: "Mobile",
-              value: "MOBILE",
-              exact: true,
-            },
-            "Would you like to receive your results via text message?": {
-              label: "Yes",
-              value: "SMS",
-              exact: false,
-            },
-            Race: {
-              label: "Other",
-              value: "other",
-              exact: true,
-            },
-            "Are you Hispanic or Latino?": {
-              label: "Prefer not to answer",
-              value: "refused",
-              exact: true,
-            },
-            "Sex assigned at birth": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-            "What's your gender identity?": {
-              label: "Female",
-              value: "female",
-              exact: true,
-            },
-          }
-        );
-        await user.click(
-          screen.queryAllByText("Save and start test", {
-            exact: false,
-          })[0]
-        );
-        expect(
-          await screen.findByText("Address validation", {
-            exact: false,
-          })
-        ).toBeInTheDocument();
-
-        const modal = screen.getByRole("dialog");
-
-        await user.click(
-          within(modal).getByLabelText("Use address as entered", {
-            exact: false,
-          })
-        );
-        await user.click(
-          within(modal).getByText("Save changes", {
-            exact: false,
-          })
-        );
-
-        expect(
-          await screen.findByText("Testing Queue!", { exact: false })
-        ).toBeInTheDocument();
-
-        expect(
-          await screen.findByText("facility-id-001", { exact: false })
-        ).toBeInTheDocument();
-
-        expect(
-          screen.getByText("153f661f-b6ea-4711-b9ab-487b95198cce", {
-            exact: false,
-          })
-        ).toBeInTheDocument();
-      });
-    });
-  });
-
-  describe("when attempting to create an existing patient ", () => {
-    const renderWithUser = (mocks: any[]) => ({
-      user: userEvent.setup(),
-      ...render(
-        <Provider store={store}>
-          <MockedProvider mocks={mocks} addTypename={false}>
-            <RouterWithFacility>
-              <Route element={<AddPatient />} path={"/add-patient/"} />
-              <Route path={"/patients"} element={<p>Patients!</p>} />
-            </RouterWithFacility>
-          </MockedProvider>
-        </Provider>
-      ),
-    });
-    it("does not open modal if no patient with matching data exists", async () => {
-      let patientExistsMock = jest.fn();
-      const mocks = [
-        {
-          request: {
-            query: PATIENT_EXISTS,
-            variables: {
-              firstName: "Alice",
-              lastName: "Hamilton",
-              birthDate: "1970-09-22",
-              facilityId: mockFacilityID,
-            },
-          },
-          result: () => {
-            patientExistsMock();
-            return {
-              data: {
-                patientExistsWithoutZip: false,
-              },
-            };
-          },
-        },
-      ];
-
-      const { user } = renderWithUser(mocks);
-
-      await fillOutForm(
-        user,
-        {
-          "First Name": "Alice",
-          "Last Name": "Hamilton",
-        },
-        { Facility: mockFacilityID },
-        {}
-      );
-
-      // The duplicate patient check is triggered on-blur from one of the identifying data fields
-      await user.type(
-        screen.getByLabelText("Date of birth", { exact: false }),
-        "1970-09-22"
-      );
-      await user.tab();
-
-      await waitFor(() => {
-        expect(patientExistsMock).toHaveBeenCalledTimes(1);
-      });
-
-      expect(
-        screen.queryByText("This patient is already registered", {
-          exact: false,
-        })
-      ).not.toBeInTheDocument();
-    });
-
-    it("displays modal when all identifying data fields have been entered with an existing patient's data", async () => {
-      const mocks = [
-        {
-          request: {
-            query: PATIENT_EXISTS,
-            variables: {
-              firstName: "Alice",
-              lastName: "Hamilton",
-              birthDate: "1970-09-22",
-              facilityId: mockFacilityID,
-            },
-          },
-          result: {
-            data: {
-              patientExistsWithoutZip: true,
-            },
-          },
-        },
-      ];
-
-      const { user } = renderWithUser(mocks);
-
-      await fillOutForm(
-        user,
-        {
-          "First Name": "Alice",
-          "Last Name": "Hamilton",
-        },
-        { Facility: mockFacilityID },
-        {}
-      );
-
-      // The duplicate patient check is triggered on-blur from one of the identifying data fields
-      await user.type(
-        screen.getByLabelText("Date of birth", { exact: false }),
-        "1970-09-22"
-      );
-      await user.tab();
-
-      expect(
-        await screen.findByText("This patient is already registered", {
-          exact: false,
-        })
-      ).toBeInTheDocument();
     });
   });
 });
